# Use batch APIs to create Array and Hash objects