--- a/ext/json/ext/parser/extconf.rb
+++ b/ext/json/ext/parser/extconf.rb
@@ -2,7 +2,10 @@
 require 'mkmf'
 
 have_func("rb_enc_interned_str", "ruby.h") # RUBY_VERSION >= 3.0
-have_func("rb_gc_mark_locations") # Missing on TruffleRuby
+have_func("rb_hash_new_capa", "ruby.h") # RUBY_VERSION >= 3.2
+have_func("rb_gc_mark_locations", "ruby.h") # Missing on TruffleRuby
+have_func("rb_hash_bulk_insert", "ruby.h") # Missing on TruffleRuby
+
 append_cflags("-std=c99")
 
 create_makefile 'json/ext/parser'